% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate.R
\name{add_approx_unique_id}
\alias{add_approx_unique_id}
\title{add_approx_unique_id
Add an approximate unique id to rows for a given column}
\usage{
add_approx_unique_id(d, col, id = NULL, max_dist = 0.1, uidName = NULL,
  uidPrefix = NULL)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
add_approx_unique_id
Add an approximate unique id to rows for a given column
}
\examples{
add(1, 1)
}

