% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{add_unique_id}
\alias{add_unique_id}
\title{add_unique_id
Find possible duplicates}
\usage{
add_unique_id(d, col, uidName = NULL, uidPrefix = NULL)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
add_unique_id
Find possible duplicates
}
\examples{
add(1, 1)
}
