% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{approx_dup}
\alias{approx_dup}
\alias{clean_col_categories}
\title{clean_col
Find possible duplicates}
\usage{
clean_col_categories(d, col, dic = NULL, remove = TRUE)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
clean_col
Find possible duplicates
}
\examples{
add(1, 1)
}
