% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{create_idcols}
\alias{create_idcols}
\title{create_idcols
Find possible duplicates}
\usage{
create_idcols(d, id_cols, id = NULL, idName = NULL, keepCols = FALSE,
  noAccents = TRUE, lowerCase = TRUE)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
create_idcols
Find possible duplicates
}
\examples{
add(1, 1)
}

