% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{exclusive_ids}
\alias{exclusive_ids}
\title{exclusive_ids
Find if two ids are exclusive}
\usage{
exclusive_ids(d, ids, .row_id = NULL, keepCols = FALSE)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
exclusive_ids
Find if two ids are exclusive
}
\examples{
add(1, 1)
}

