% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate.R
\name{new_or_dup}
\alias{new_or_dup}
\title{new_or_dup
Find if new or duplicate records}
\usage{
new_or_dup(dnew, dold, id_cols, id = NULL, max_dist = 0.1)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
new_or_dup
Find if new or duplicate records
}
\examples{
add(1, 1)
}

