% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R, R/deduplicate.R
\name{approx_dup}
\alias{approx_dup}
\alias{clean_col_categories}
\alias{get_approx_dup_ids}
\alias{get_approx_dups}
\title{clean_col
Find possible duplicates}
\usage{
clean_col_categories(d, col, dic = NULL, remove = TRUE)

get_approx_dups(d, id_cols, id = NULL, max_dist = 0.1)

get_approx_dup_ids(d, id_cols, id = NULL, max_dist = 0.1, asList = FALSE)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}

\item{x}{A number.}

\item{y}{A number.}

\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.

The sum of \code{x} and \code{y}.

The sum of \code{x} and \code{y}.
}
\description{
clean_col
Find possible duplicates

get_approx_dups
Find possible duplicates

get_approx_dup_ids
Find possible duplicates
}
\examples{
add(1, 1)
add(1, 1)
add(1, 1)
}

